storagePath 의존성 분리, MIME 우회 차단, 디코드 실패 fail-fast
- buildImageStoragePath가 sanitizeHtml.ts를 통해 dompurify와 document에 transitively 의존해 번들이 무거워지고 비-에디터 컨텍스트 재사용이 막혔음. formatDate를 DOM 독립 모듈로 분리
- 확장자 화이트리스트가 MIME에 무관하게 valid를 반환해 notes.txt를 notes.jpg로 위장하면 통과되던 우회 경로가 있었음. 확장자 매칭은 MIME이 비었거나 application/octet-stream일 때만 적용
- resize/decode 실패 시 원본 바이트를 그대로 업로드해 broken <img>가 게시되던 contract 결함을 fail-fast로 차단. 캔버스가 못 읽는 파일은 브라우저도 못 렌더링하므로 업로드 의미 없음
