+++ apps/web/src/post/utils/sanitizeHtml.ts
@@ -58,19 +58,3 @@ export function isValidHttpUrl(url: string): boolean {
   }
 }
 
-/**
- * Format date for Firebase storage path: YYYYMMDD folder + HHMMSS prefix.
- */
-export function formatDate(date: Date): { dateFolder: string; timePrefix: string } {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  const hours = String(date.getHours()).padStart(2, '0');
-  const minutes = String(date.getMinutes()).padStart(2, '0');
-  const seconds = String(date.getSeconds()).padStart(2, '0');
-
-  return {
-    dateFolder: `${year}${month}${day}`,
-    timePrefix: `${hours}${minutes}${seconds}`
-  };
-}
